+++ connectors/src/resources/db/schema/index.ts
@@ -1,2 +1,3 @@
-export * from "./confluence";
 export * from "./connectors";
+export * from "./github";
+export * from "./schema";
