+++ connectors/src/resources/db/schema/schema.ts
@@ -1,7 +1,5 @@
-import { sql } from "drizzle-orm";
 import {
   boolean,
-  foreignKey,
   index,
   integer,
   jsonb,
@@ -13,52 +11,9 @@ import {
   varchar,
 } from "drizzle-orm/pg-core";
 
-export const connectors = pgTable(
-  "connectors",
-  {
-    id: serial("id").primaryKey().notNull(),
-    createdAt: timestamp("createdAt", {
-      withTimezone: true,
-      mode: "string",
-    }).notNull(),
-    updatedAt: timestamp("updatedAt", {
-      withTimezone: true,
-      mode: "string",
-    }).notNull(),
-    type: varchar("type", { length: 255 }).notNull(),
-    connectionId: varchar("connectionId", { length: 255 }).notNull(),
-    workspaceApiKey: varchar("workspaceAPIKey", { length: 255 }).notNull(),
-    workspaceId: varchar("workspaceId", { length: 255 }).notNull(),
-    dataSourceName: varchar("dataSourceName", { length: 255 }).notNull(),
-    lastSyncStatus: varchar("lastSyncStatus", { length: 255 }),
-    errorType: varchar("errorType", { length: 255 }),
-    lastSyncStartTime: timestamp("lastSyncStartTime", {
-      withTimezone: true,
-      mode: "string",
-    }),
-    lastSyncFinishTime: timestamp("lastSyncFinishTime", {
-      withTimezone: true,
-      mode: "string",
-    }),
-    lastSyncSuccessfulTime: timestamp("lastSyncSuccessfulTime", {
-      withTimezone: true,
-      mode: "string",
-    }),
-    firstSuccessfulSyncTime: timestamp("firstSuccessfulSyncTime", {
-      withTimezone: true,
-      mode: "string",
-    }),
-    firstSyncProgress: varchar("firstSyncProgress", { length: 255 }),
-    lastGcTime: timestamp("lastGCTime", { withTimezone: true, mode: "string" }),
-  },
-  (table) => {
-    return {
-      workspaceIdDataSourceName: uniqueIndex(
-        "connectors_workspace_id_data_source_name"
-      ).on(table.workspaceId, table.dataSourceName),
-    };
-  }
-);
+import { connectors } from "@connectors/resources/db/schema/connectors";
+
+// /!\ The schema files SHOULD NOT contain any runtime logic besides defining your DB schema.
 
 export const slackChannels = pgTable(
   "slack_channels",
@@ -1117,87 +1072,6 @@ export const confluencePages = pgTable(
       connectorIdLastVisitedAt: index(
         "confluence_pages_connector_id_last_visited_at"
       ).on(table.connectorId, table.lastVisitedAt),
-    };
-  }
-);
-
-export const githubCodeRepositories = pgTable(
-  "github_code_repositories",
-  {
-    id: serial("id").primaryKey().notNull(),
-    createdAt: timestamp("createdAt", {
-      withTimezone: true,
-      mode: "string",
-    }).notNull(),
-    updatedAt: timestamp("updatedAt", {
-      withTimezone: true,
-      mode: "string",
-    }).notNull(),
-    lastSeenAt: timestamp("lastSeenAt", {
-      withTimezone: true,
-      mode: "string",
-    }).notNull(),
-    repoId: varchar("repoId", { length: 255 }).notNull(),
-    repoLogin: varchar("repoLogin", { length: 255 }).notNull(),
-    repoName: varchar("repoName", { length: 255 }).notNull(),
-    sourceUrl: varchar("sourceUrl", { length: 255 }).notNull(),
-    connectorId: integer("connectorId").references(() => connectors.id, {
-      onDelete: "set null",
-      onUpdate: "cascade",
-    }),
-    codeUpdatedAt: timestamp("codeUpdatedAt", {
-      withTimezone: true,
-      mode: "string",
-    }).notNull(),
-  },
-  (table) => {
-    return {
-      connectorIdRepoId: uniqueIndex(
-        "github_code_repositories_connector_id_repo_id"
-      ).on(table.repoId, table.connectorId),
-    };
-  }
-);
-
-export const githubCodeFiles = pgTable(
-  "github_code_files",
-  {
-    id: serial("id").primaryKey().notNull(),
-    createdAt: timestamp("createdAt", {
-      withTimezone: true,
-      mode: "string",
-    }).notNull(),
-    updatedAt: timestamp("updatedAt", {
-      withTimezone: true,
-      mode: "string",
-    }).notNull(),
-    lastSeenAt: timestamp("lastSeenAt", {
-      withTimezone: true,
-      mode: "string",
-    }).notNull(),
-    repoId: varchar("repoId", { length: 255 }).notNull(),
-    documentId: varchar("documentId", { length: 255 }).notNull(),
-    parentInternalId: varchar("parentInternalId", { length: 255 }).notNull(),
-    fileName: varchar("fileName", { length: 255 }).notNull(),
-    sourceUrl: varchar("sourceUrl", { length: 255 }).notNull(),
-    contentHash: varchar("contentHash", { length: 255 }).notNull(),
-    connectorId: integer("connectorId").references(() => connectors.id, {
-      onDelete: "set null",
-      onUpdate: "cascade",
-    }),
-    codeUpdatedAt: timestamp("codeUpdatedAt", {
-      withTimezone: true,
-      mode: "string",
-    }).notNull(),
-  },
-  (table) => {
-    return {
-      connectorIdRepoIdLastSeenAt: index(
-        "github_code_files_connector_id_repo_id_last_seen_at"
-      ).on(table.lastSeenAt, table.repoId, table.connectorId),
-      connectorIdRepoIdDocumentId: uniqueIndex(
-        "github_code_files_connector_id_repo_id_document_id"
-      ).on(table.repoId, table.documentId, table.connectorId),
     };
   }
 );
